requirements/base.txt
django
djangorestframework
gunicorn
django-shell-plus
django_extensions
django-cors-headers
django-environ
django-rest-swagger
ipdb
pymysql
pandas
django-bootstrap4
markdown
django-ckeditor